requirements.txt
mediapipe==0.10.21
opencv-python==4.11.0.86
pyserial==3.5
numpy==2.2.4
